clinicbackend/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.5
numpy==2.3.2
pillow==11.3.0
platformdirs==4.3.8
PyJWT==2.10.1
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
typing_extensions==4.14.1
tzdata==2025.2
virtualenv==20.31.2
